requirements.txt
PyYAML==3.11
feedparser==5.2.1
simplejson
pytest
